+++ sources/pooler.c
@@ -225,7 +225,7 @@ od_pooler_config_import(od_pooler_t *pooler)
 
 	od_scheme_t scheme;
 	od_scheme_init(&scheme);
-	int scheme_version;
+	uint64_t scheme_version;
 	scheme_version = od_schememgr_version_next(&instance->scheme_mgr);
 	int rc;
 	rc = od_config_load(&scheme, &instance->logger, instance->config_file,
